backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
qdrant-client
pydantic
python-docx
PyMuPDF
sentence-transformers
numpy
